implement play() metho
