+++ src/main/java/com/f5/ghostbuster/App.java
@@ -1,13 +1,16 @@
 package com.f5.ghostbuster;
 
-/**
- * Hello world!
- *
- */
+import com.f5.ghostbuster.controller.GhostBusterController;
+import com.f5.ghostbuster.models.GhostBuster;
+import com.f5.ghostbuster.views.ConsoleView;
+
 public class App 
 {
     public static void main( String[] args )
     {
-        System.out.println( "Hello World!" );
+        GhostBuster ghostBuster = new GhostBuster();
+        ConsoleView consoleView = new ConsoleView();
+        GhostBusterController ghostController = new GhostBusterController(ghostBuster, consoleView);
+        ghostController.run();
     }
 }
